second commit

+++ src/components/Model.js
@@ -40,10 +40,9 @@ const Model = ({alldata_list,device_name}) => {
 
   const device1_convert = parseInt(device1_data)
 
-
   const limitvalue = ((device1_convert-0)*(100-0))/(device1_thickness-0)+0;
   const rounded_value_device1 = limitvalue.toFixed(2);
-  const rounded_percentage_device1 = parseFloat(rounded_value_device1);
+  const rounded_percentage_device1 = parseInt(rounded_value_device1);
   const device1_thickness_int = parseInt(rounded_percentage_device1);
   const device2_convert = parseInt(device2_data)
 
@@ -51,26 +50,26 @@ const Model = ({alldata_list,device_name}) => {
   
   const device2_limitvalue2 = ((device2_convert-0)*(100-0))/(device2_thickness-0)+0;
   const rounded_value_device2 = device2_limitvalue2.toFixed(2);
-  const rounded_percentage_device2 = parseFloat(rounded_value_device2);
+  const rounded_percentage_device2 = parseInt(rounded_value_device2);
   const device2_thickness_int = parseInt(rounded_percentage_device2);
 
   const device3_convert = parseInt(device3_data)
   const device3_limitvalue2 = ((device3_convert-0)*(100-0))/(device3_thickness-0)+0;
   const rounded_value_device3 = device3_limitvalue2.toFixed(2);
-  const rounded_percentage_device3 = parseFloat(rounded_value_device3);
+  const rounded_percentage_device3 = parseInt(rounded_value_device3);
   const device3_thickness_int = parseInt(rounded_percentage_device3);
 
   const device4_convert = parseInt(device4_data)
   const device4_limitvalue2 = ((device4_convert-0)*(100-0))/(device4_thickness-0)+0;
   const rounded_value_devic4 = device4_limitvalue2.toFixed(2);
-  const rounded_percentage_device4 = parseFloat(rounded_value_devic4);
+  const rounded_percentage_device4 = parseInt(rounded_value_devic4);
   const device4_thickness_int = parseInt(rounded_percentage_device4);
 
 
   const device5_convert = parseInt(device5_data)
   const device5_limitvalue2 = ((device5_convert-0)*(100-0))/(device5_thickness-0)+0;
   const rounded_value_devic5 = device5_limitvalue2.toFixed(2);
-  const rounded_percentage_device5 = parseFloat(rounded_value_devic5);
+  const rounded_percentage_device5 = parseInt(rounded_value_devic5);
   const device5_thickness_int = parseInt(rounded_percentage_device5);
